+++ tests/subcutaneous/theia_world_test.py
@@ -1,0 +1,15 @@
+"""Tests of creating, saving and loading Thea world objects."""
+
+# from thea.theia_world import TeiaWorld
+
+# def test_world_loading():
+
+#     original_tw = TheiaWorld()
+#     original_tw.new()
+#     original_tw.save()
+
+#     loaded_tw = TheiaWorld()
+#     loaded_tw.load()
+
+#     # How to make this work?
+#     assert original_tw == loaded_tw